feat(lib): make a new 'lib' submodule to hold assorted utilities
