+++ caput/lib/__init__.py
@@ -1,0 +1,13 @@
+"""caput.lib.
+
+Assorted utilities that don't belong anywhere else.
+"""
+
+from . import (
+    array_utils as array_utils,
+    cache as cache,
+    misc as misc,
+    pfb as pfb,
+    profile as profile,
+    random as random,
+)
