+++ caput/lib/random.py
